% Generated by build script
\name{epa_facility}
\alias{epa_facility}
\title{EPA facility registry}
\usage{
epa_facility()
}
\arguments{
\item{state}{2-letter state code}
\item{city}{City name filter}
\item{zip}{ZIP code filter}
\item{max_rows}{Max rows (default 1000)}
}
\value{tibble: registry_id, primary_name, city_name, state_code,}
\description{EPA facility registry}
