% Generated by build script
\name{epa_ghg_facility}
\alias{epa_ghg_facility}
\title{Greenhouse gas reporting facilities (GHGRP)}
\usage{
epa_ghg_facility()
}
\arguments{
\item{state}{2-letter state abbreviation}
\item{max_rows}{Max rows (default 1000)}
}
\value{tibble: facility_id, latitude, longitude, city, state, zip,}
\description{Greenhouse gas reporting facilities (GHGRP)}
