% Generated by build script
\name{epa_tri_facility}
\alias{epa_tri_facility}
\title{TRI facility information}
\usage{
epa_tri_facility()
}
\arguments{
\item{state}{2-letter state abbreviation (e.g. "CA", "TX")}
\item{city}{City name filter}
\item{zip}{ZIP code filter}
\item{max_rows}{Max rows (default 1000)}
}
\value{tibble: tri_facility_id, facility_name, street_address, city_name,}
\description{TRI facility information}
