% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/epa.R
\name{epa_tri_release}
\alias{epa_tri_release}
\title{TRI chemical release quantities}
\usage{
epa_tri_release(state = NULL, year = NULL, chemical = NULL, max_rows = 5000)
}
\arguments{
\item{state}{2-letter state abbreviation}

\item{year}{Reporting year (e.g. "2022")}

\item{chemical}{Chemical name filter}

\item{max_rows}{Max rows (default 5000)}
}
\value{
tibble: tri_facility_id, chemical_name, unit_of_measure,
total_releases, on_site_release_total, off_site_release_total, ...
}
\description{
Chemical release data from TRI-reporting facilities.
}
