% Generated by build script
\name{epa_water_permit}
\alias{epa_water_permit}
\title{Water discharge permits (PCS/NPDES)}
\usage{
epa_water_permit()
}
\arguments{
\item{state}{2-letter state code}
\item{city}{City name filter}
\item{max_rows}{Max rows (default 1000)}
}
\value{tibble with permit and facility information}
\description{Water discharge permits (PCS/NPDES)}
