% Generated by build script
\name{epa_tri_release}
\alias{epa_tri_release}
\title{TRI chemical release quantities}
\usage{
epa_tri_release()
}
\arguments{
\item{state}{2-letter state abbreviation}
\item{year}{Reporting year (e.g. "2022")}
\item{chemical}{Chemical name filter}
\item{max_rows}{Max rows (default 5000)}
}
\value{tibble: tri_facility_id, chemical_name, unit_of_measure,}
\description{TRI chemical release quantities}
